For issue #1: negative offsets and out-of-bounds origins
Add some test cases, and then resolve the issue with a test in _quadrant().  It doesn't really make sense to ask for a field of view from an origin outside of the map, because nothing is known about the cells or walls out there.

But if someone accidentally asks for it, we shouldn't have an infinite loop.  So here we just treat that outside area as not visible.